updated endpoints

+++ sanic-backend/app.py
@@ -7,7 +7,7 @@ from tempfile import TemporaryDirectory
 
 import psycopg2
 from sanic import Sanic
-from sanic.response import json as sanic_json, text
+from sanic.response import json as sanic_json, text as sanic_text
 from sanic_cors.extension import CORS
 from sanic_ext import Extend, cors
 from synthetic.pdf.parser import parse_pdf
@@ -88,31 +88,22 @@ def run_synthsizer(request, orgId):
             return sanic_json({"received": False,
                            "message": "GT not handled correctly"})
 
-
-      
-
-        for pdf in pdf_collection:
-            print(pdf.name)
-        
-        for gt in gt_collection:
-            print(gt.name)
-
         for i in range(len(pdf_collection)):
             # add pdf to synthesized table
             #data = b64encode(pdf_collection[i].read_bytes()).decode('utf-8')
             cursor.execute("INSERT INTO synthesized (pdf, gt, orgID) VALUES (%s, %s, %s)",
                            (pdf_collection[i].read_bytes(), gt_collection[i].read_text(), orgId))
 
-        return text("Synthesized")
+        return sanic_text(status)
 
 
 # GET request to get the certain row from the database
-@app.route('/documents/<documentId>', methods=['GET'])
+@app.route('/documents/<orgId>/<documentId>', methods=['GET'])
 @cors(allow_methods="GET")
-def fetch_document(request, documentId):
+def fetch_document(request, orgId, documentId):
 
     logging.info("Fetching document")
-    cursor.execute("SELECT * FROM synthesized WHERE id = %s;", (documentId,))
+    cursor.execute("SELECT * FROM synthesized WHERE id = %s AND orgId = %s", (documentId,orgId))
     row = cursor.fetchone()
     json_statement = create_statement(row[1], row[2], row[3])
     
@@ -120,11 +111,11 @@ def fetch_document(request, documentId):
 
 
 # DELETE request that deletes the row with the given id
-@app.route('/documents/<documentId>', methods=['DELETE'])
+@app.route('/documents/<orgId>/<documentId>', methods=['DELETE'])
 @cors(allow_methods="DELETE")
-def delete_document(request, documentId):
-    cursor.execute("DELETE FROM synthesized WHERE id = %s;", (documentId,))
-    return text("Deleted")
+def delete_document(request, orgId, documentId):
+    cursor.execute("DELETE FROM synthesized WHERE id = %s AND orgId = %s;", (documentId,orgId))
+    return sanic_text("Deleted")
 
 
 # GET request that returns the all the rows in the database
@@ -141,19 +132,19 @@ def all_documents(request):
 
     return sanic_json({"received": True, "message": json_statment})
 
-# # GET request that returns the all the rows accoring to the orgID
-# @app.route('/documents/<orgId>', methods=['GET'])
-# @cors(allow_methods="GET")
-# def all_documents_org(request, orgId):
-#     cursor.execute("SELECT * FROM synthesized WHERE orgID = %s;", (orgId,))
-#     data = cursor.fetchall()
-#     json_statment = {}
-#     for data_pair in data:
-#         json_statment[data_pair[0]] = create_statement(data_pair[1],
-#                                                        data_pair[2],
-#                                                        data_pair[3])
+# GET request that returns the all the rows accoring to the orgID
+@app.route('/documents/<orgId>', methods=['GET'])
+@cors(allow_methods="GET")
+def all_documents_org(request, orgId):
+    cursor.execute("SELECT * FROM synthesized WHERE orgID = %s;", (orgId,))
+    data = cursor.fetchall()
+    json_statment = {}
+    for data_pair in data:
+        json_statment[data_pair[0]] = create_statement(data_pair[1],
+                                                       data_pair[2],
+                                                       data_pair[3])
 
-#     return sanic_json({"received": True, "message": json_statment})
+    return sanic_json({"received": True, "message": json_statment})
 
 # Synthesizes the pdf and gt
 def synthesize_document(
